10-RAG-document-storage/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Vector Database
langchain-chroma>=0.1.0
chromadb>=0.4.0

# Document Loaders (for examples)
pypdf>=3.0.0  # For PDF loading
unstructured>=0.10.0  # For HTML and other unstructured formats

# Hugging Face Support (optional, for embeddings)
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0

# Optional: For environment variable management
python-dotenv>=1.0.0

